Replace system test with an integration test.
